fix cycle dependency between api and client
